Forwardに対するBulletのデータ作成。完了

+++ Assets/Sasaki/Scripts/Data/Bullet/BulletParam.cs
@@ -1,6 +1,6 @@
 public class BulletParam
 {
-    public float WayCount;
+    public int WayCount;
     public float Angle;
     public float Blur;
 }
